backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
appwrite
qdrant-client
redis
rq
openai
pydantic
pydantic-settings
python-dotenv
pytest
httpx
google-generativeai
pypdf
python-docx
tenacity
sentence-transformers
